+++ src/App.tsx
@@ -10,6 +10,7 @@ import { CommandPalette } from './components/CommandPalette';
 import { StatusBar } from './components/StatusBar';
 import { Toast } from './components/Toast';
 import { CheatSheet } from './components/CheatSheet';
+import { UpdateChecker } from './components/UpdateChecker';
 import { SettingsButton } from './components/SettingsPanel';
 import { applyTheme } from '@/lib/themes';
 import { applyReducedMotion } from '@/lib/sounds';
@@ -274,6 +275,7 @@ export function App() {
       <Composer />
       <CommandPalette />
       <CheatSheet />
+      <UpdateChecker />
       <Toast />
       {showOnboarding && <Onboarding onDone={() => setShowOnboarding(false)} />}
     </div>